+++ p2p_wallet/Scenes/Main/History/HistoryItem.swift
@@ -27,7 +27,7 @@ enum HistoryItem: Hashable {
         case .parsedTransaction(let transaction):
             return transaction.blockTime
         case .sellTransaction(let transaction):
-            return transaction.createdAt
+            return nil
         }
     }
 }
